target: Set default access method for PROC target to SBEFIFO
Set the default hardware access method for all processor (PROC) targets
to SBEFIFO, which will be used for normal chip communication.

In scenarios where the BMC fails to communicate with the SBE (e.g.,
SBE not responding or FIFO timeout).

BMC will switch to the direct FSI access method to retrieve debug and
recovery data.

Change-Id: Ie9fabe4fe52bc857cf3fe15d05dcb566736b66cb